Moving analytics hooks
